+++ GameFramework/Include/Object/Player.h
@@ -76,7 +76,7 @@ public:
 	virtual void Collision(float DeltaTime);
 	virtual void Render(HDC hDC);
 	virtual CPlayer *Clone();
-	virtual float SetDamage(int Damage);
+	virtual int SetDamage(int Damage);
 // Hp,Mp,Stemina (Ability)
 public :
 	void AbilityUpdate(float DeltaTime);
