+++ GameFramework/Include/Object/PlayerClone.h
@@ -48,7 +48,7 @@ public :
 	virtual void PostUpdate(float DeltaTime);
 	virtual void Collision(float DeltaTime);
 	virtual void Render(HDC hDC);
-	virtual float SetDamage(float Damage);
+	virtual int SetDamage(int Damage);
 	virtual void CharacterDestroy();
 public:
 	void Fire();
